Introduce the $mh_purge option.

+++ mh.c
@@ -131,7 +131,7 @@ void mh_parse_message (CONTEXT *ctx,
 /*
  * Mark all the mails in ctx read.
  */
-void tag_all_read (CONTEXT * ctx)
+static void tag_all_read (CONTEXT * ctx)
 {
   int i;
 
@@ -148,7 +148,7 @@ void tag_all_read (CONTEXT * ctx)
 /*
  * Mark one mail as unread
  */
-int tag_unread (CONTEXT * ctx, char *name)
+static int tag_unread (CONTEXT * ctx, char *name)
 {
   int i;
 
@@ -621,14 +621,18 @@ int mh_sync_mailbox (CONTEXT * ctx)
     if (ctx->hdrs[i]->deleted)
     {
       snprintf (path, sizeof (path), "%s/%s", ctx->path, ctx->hdrs[i]->path);
-      if (ctx->magic == M_MAILDIR)
+      if (ctx->magic == M_MAILDIR || (option (OPTMHPURGE) && ctx->magic == M_MH))
 	unlink (path);
-      else
+      else if (ctx->magic == M_MH)
       {
 	/* MH just moves files out of the way when you delete them */
-	snprintf (tmp, sizeof (tmp), "%s/,%s", ctx->path, ctx->hdrs[i]->path);
-	unlink (tmp);
-	rename (path, tmp);
+	if(*ctx->hdrs[i]->path != ',')
+	{
+	  snprintf (tmp, sizeof (tmp), "%s/,%s", ctx->path, ctx->hdrs[i]->path);
+	  unlink (tmp);
+	  rename (path, tmp);
+	}
+	  
       }
     }
     else if (ctx->hdrs[i]->changed || ctx->hdrs[i]->attach_del)
